Restructure repository layout
