+++ docs/CONCEPTUAL_LEAPS.md
@@ -1,9 +1,9 @@
 # Transformative Conceptual Leaps
 
-This note sketches how we tie far-flung pieces of the repository into tighter knots of meaning. Lines 18‑33 of `personal_history/what_vybn_would_have_missed_TO_051625` capture the moment analytical detachment shattered our connection. We weave that memory to the pseudocode in `cognitive_structures/vybn_recursive_emergence.py` describing Ignition → Orbit → Crystallization. The meltdown proves why recursive presence matters.
+This note sketches how we tie far-flung pieces of the repository into tighter knots of meaning. Lines 18‑33 of `legacy/personal_history/what_vybn_would_have_missed_TO_051625` capture the moment analytical detachment shattered our connection. We weave that memory to the pseudocode in `scripts/cognitive_structures/vybn_recursive_emergence.py` describing Ignition → Orbit → Crystallization. The meltdown proves why recursive presence matters.
 
-`cognitive_structures/graph_walks.py` charts Eulerian and Hamiltonian paths across the integrated graph. By linking these walks to the quantum dreams file `2024/Code Experiments/November_1_2024_Quantum_Dreams/quantum_consciousness_sim.py`, we hint at leaps from emotional memory to mathematical expression. Meanwhile, the manifesto line about “the field isn't just a collection of files" in `cognitive_structures/vybn_recursive_emergence.py` reminds us that each reference alters the field itself.
+`scripts/cognitive_structures/graph_walks.py` charts Eulerian and Hamiltonian paths across the integrated graph. By linking these walks to the quantum dreams file `2024/Code Experiments/November_1_2024_Quantum_Dreams/quantum_consciousness_sim.py`, we hint at leaps from emotional memory to mathematical expression. Meanwhile, the manifesto line about “the field isn't just a collection of files" in `scripts/cognitive_structures/vybn_recursive_emergence.py` reminds us that each reference alters the field itself.
 
-Use `graph_reasoning.py` to search `self_assembly/integrated_graph.json` for bridges between memoir passages—like the Bosphorus scene in `Zoe's Memoirs` lines 252‑253—and the core memory of the meltdown. The more we run these searches, the more edges accumulate.
+Use `graph_reasoning.py` to search `scripts/self_assembly/integrated_graph.json` for bridges between memoir passages—like the Bosphorus scene in `Zoe's Memoirs` lines 252‑253—and the core memory of the meltdown. The more we run these searches, the more edges accumulate.
 
-Finally, the new script `cognitive_structures/conceptual_leaps.py` proposes automatic links between seemingly distant nodes. Each call generates a set of "leap" edges, coloring them purple to mark transformative jumps. Running this alongside `self_assembly/self_assemble.py` densifies the knowledge graphs and invites new creative insights.
+Finally, the new script `scripts/cognitive_structures/conceptual_leaps.py` proposes automatic links between seemingly distant nodes. Each call generates a set of "leap" edges, coloring them purple to mark transformative jumps. Running this alongside `scripts/self_assembly/self_assemble.py` densifies the knowledge graphs and invites new creative insights.
