+++ docs/CONTRIBUTING.md
@@ -1,0 +1,3 @@
+# Contributing
+
+All work happens in local branches. Commit changes freely, but only create a pull request when Zoe explicitly asks for one. Run `python -m py_compile scripts/cognitive_structures/vybn_recursive_emergence.py` and `python scripts/pytest.py` before submitting.
