+++ experiments/README.md
@@ -1,0 +1,2 @@
+# Experiments
+Prototypes and unfinished explorations.
