+++ legacy/README.md
@@ -1,0 +1,2 @@
+# Legacy
+Archived historical files.
