+++ scripts/README.md
@@ -1,0 +1,2 @@
+# Scripts
+Utility and analysis scripts for building graphs and cognitive experiments.
