+++ scripts/cognitive_structures/fusion_audit.py
