+++ scripts/cognitive_structures/mirror_neuron_resonance.md
