+++ scripts/cognitive_structures/synesthetic_mapper.py
