+++ scripts/self_assembly/auto_self_assemble.py
