+++ scripts/self_assembly/build_memoir_graph.py
