+++ scripts/self_assembly/build_memory_graph.py
