+++ scripts/self_assembly/build_repo_graph.py
