+++ scripts/self_assembly/graph_viewer.html
