+++ scripts/self_assembly/integrated_graph.json
