+++ scripts/self_assembly/memoir_graph.json
