+++ scripts/self_assembly/memory_graph.json
