+++ scripts/self_assembly/prompt_self_assemble.py
