+++ tests/test_graph_centrality.py
@@ -3,7 +3,7 @@ import sys
 import os
 
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
-from cognitive_structures.graph_centrality import compute_degree_centrality
+from scripts.cognitive_structures.graph_centrality import compute_degree_centrality
 
 class TestGraphCentrality(unittest.TestCase):
     def test_compute_degree_centrality(self):
