+++ tests/test_vybn_compile.py
@@ -4,7 +4,7 @@ import unittest
 
 class TestCompile(unittest.TestCase):
     def test_vybn_recursive_emergence_compiles(self):
-        path = os.path.join('cognitive_structures', 'vybn_recursive_emergence.py')
+        path = os.path.join('scripts', 'cognitive_structures', 'vybn_recursive_emergence.py')
         py_compile.compile(path, doraise=True)
 
 if __name__ == '__main__':
